+++ src/test/run-make/libtest-json/f.rs
@@ -29,3 +29,4 @@ fn c() {
 fn d() {
     assert!(false);
 }
+
